SubmissionAdjustments

+++ backend/CampusCore.API/Services/IIssueService.cs
@@ -8,12 +8,12 @@ namespace CampusCore.API.Services
     public interface IIssueService
     {
         Task<ResponseManager> CreateIssueAsync(IssueAddViewModel model);
-        Task<ResponseManager> ViewIssueListAsync(IntIdViewModel model);
-        Task<ResponseManager> ViewIssueListOpenAsync();
-        Task<ResponseManager> IssueGetByIdAsync(GetByIdModel model);
-        Task<ResponseManager> DeleteIssueAsync(IssueDeleteModel model);
+        Task<ResponseManager> GetAllByUserAsync(IssueGetAllModel model); //get issue that concerns a user (can be student or faculty, the one who needs to resolve the issue) not the one who opened the issue
+        Task<ResponseManager> GetAllBySubmissionAsync(IssueGetAllModel model);
+        Task<ResponseManager> IssueGetByIdAsync(IntIdViewModel model);
+        Task<ResponseManager> DeleteIssueAsync(IntIdViewModel model);
         Task<ResponseManager> UpdateIssueAsync(IssueUpdateViewModel model);
-        Task<ResponseManager> SearchIssueAsync(IssueSearchViewModel model);
+       // Task<ResponseManager> SearchIssueAsync(IssueSearchViewModel model); 
 
     }
 
@@ -44,13 +44,34 @@ namespace CampusCore.API.Services
             _context.Issues.Add(issue);
             var result = await _context.SaveChangesAsync();
 
+            
+            
             if (result > 0)
             {
-                return new ResponseManager
+                _context.SubmissionIssues.Add(new SubmissionIssue
                 {
-                    Message = "Issue created successfully!",
-                    IsSuccess = true
-                };
+                    IssueId = issue.Id,
+                    SubmissionId = model.SubmissionId
+                });
+                var res = await _context.SaveChangesAsync();
+                if(res > 0)
+                {
+                    return new ResponseManager
+                    {
+                        Message = "Issue created successfully!",
+                        IsSuccess = true
+                    };
+                }
+                else
+                {
+                    return new ErrorResponseManager
+                    {
+                        Message = "Issue is created but failed to associate with submission",
+                        IsSuccess = false,
+                        Errors = new List<string>() { "Error adding issue in DB" }
+                    };
+                }
+                
 
             }
 
@@ -68,64 +89,213 @@ namespace CampusCore.API.Services
 
 
         }
-        public async Task<ResponseManager> SearchIssueAsync(IssueSearchViewModel model)
+        //public async Task<ResponseManager> SearchIssueAsync(IssueSearchViewModel model)
+        //{
+        //    string searchKey = model.SearchIssue;
+
+        //    try
+        //    {
+
+        //        var searchResults = await _context.Issues
+        //            .Where(oc => EF.Functions.Like(oc.Name, $"%{model.SearchIssue}%"))
+        //            .ToListAsync();
+
+
+
+        //        return new DataResponseManager
+        //        {
+        //            IsSuccess = true,
+        //            Message = "Searched issues retrieved successfully",
+        //            Data = searchResults
+        //        };
+        //    }
+        //    catch (Exception ex)
+        //    {
+        //        return new ErrorResponseManager
+        //        {
+        //            IsSuccess = false,
+        //            Message = "An error occurred while fetching searched issues",
+        //            Errors = new List<string> { ex.Message }
+        //        };
+        //    }
+        //}
+
+        public async Task<ResponseManager> GetAllBySubmissionAsync(IssueGetAllModel model)
         {
-            string searchKey = model.SearchIssue;
+            var submissionId = model.SubmissionId;
 
-            try
+            if(model.Filter == "open")
             {
-
-                var searchResults = await _context.Issues
-                    .Where(oc => EF.Functions.Like(oc.Name, $"%{model.SearchIssue}%"))
-                    .ToListAsync();
-
-
-
-                return new DataResponseManager
+                try
                 {
-                    IsSuccess = true,
-                    Message = "Searched issues retrieved successfully",
-                    Data = searchResults
-                };
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.SubmissionId == submissionId)
+                                                .Select(si => si.Issue)
+                                                .Where(i=> i.Status == "open")
+                                                .ToListAsync();
+
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
+                {
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
             }
-            catch (Exception ex)
+            if (model.Filter == "close")
             {
-                return new ErrorResponseManager
+                try
                 {
-                    IsSuccess = false,
-                    Message = "An error occurred while fetching searched issues",
-                    Errors = new List<string> { ex.Message }
-                };
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.SubmissionId == submissionId)
+                                                .Select(si => si.Issue)
+                                                .Where(i => i.Status == "closed")
+                                                .ToListAsync();
+
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
+                {
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
             }
+            else
+            {
+                try
+                {
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.SubmissionId == submissionId)
+                                                .Select(si => si.Issue)
+                                                .ToListAsync();
+
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
+                {
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
+            }
+            
         }
 
-        public async Task<ResponseManager> ViewIssueListAsync(IntIdViewModel model)
+        public async Task<ResponseManager> GetAllByUserAsync(IssueGetAllModel model)
         {
-            var submissionId = model.Id;
+            var userId = model.UserId;
+            var userGroupID = _context.StudentGroups
+                                      .Where(sg => sg.StudentId == userId)
+                                      .Select(sg=> sg.GroupId)
+                                      .FirstOrDefault();
 
-            try
+            if (model.Filter == "open")
             {
-                var result = await _context.SubmissionIssues
-                                            .Where(si => si.SubmissionId == submissionId)
-                                            .Select(si=> si.Issue)
-                                            .ToListAsync();
+                try
+                {
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.Submission.GroupId == userGroupID || si.Submission.SubmitterId == userId)
+                                                .Select(si => si.Issue)
+                                                .Where(i => i.Status == "open")
+                                                .ToListAsync();
 
-                return new DataResponseManager
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
                 {
-                    IsSuccess = true,
-                    Message = "Issues retrieved successfully",
-                    Data = result
-                };
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
             }
-            catch (Exception ex)
+            if (model.Filter == "close")
             {
-                return new ErrorResponseManager
+                try
                 {
-                    IsSuccess = false,
-                    Message = "An error occurred while fetching issues",
-                    Errors = new List<string> { ex.Message }
-                };
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.Submission.GroupId == userGroupID || si.Submission.SubmitterId == userId)
+                                                .Select(si => si.Issue)
+                                                .Where(i => i.Status == "closed")
+                                                .ToListAsync();
+
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
+                {
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
             }
+            else
+            {
+                try
+                {
+                    var result = await _context.SubmissionIssues
+                                                .Where(si => si.Submission.GroupId == userGroupID || si.Submission.SubmitterId == userId)
+                                                .Select(si => si.Issue)
+                                                .ToListAsync();
+
+                    return new DataResponseManager
+                    {
+                        IsSuccess = true,
+                        Message = "Issues retrieved successfully",
+                        Data = result
+                    };
+                }
+                catch (Exception ex)
+                {
+                    return new ErrorResponseManager
+                    {
+                        IsSuccess = false,
+                        Message = "An error occurred while fetching issues",
+                        Errors = new List<string> { ex.Message }
+                    };
+                }
+            }
+
         }
 
         public async Task<ResponseManager> ViewIssueListOpenAsync()
@@ -157,7 +327,7 @@ namespace CampusCore.API.Services
         }
 
 
-        public async Task<ResponseManager> IssueGetByIdAsync(GetByIdModel model)
+        public async Task<ResponseManager> IssueGetByIdAsync(IntIdViewModel model)
         {
 
 
@@ -187,9 +357,9 @@ namespace CampusCore.API.Services
         // Add a default return statement or throw an exception here.
 
 
+        
 
-
-        public async Task<ResponseManager> DeleteIssueAsync(IssueDeleteModel model)
+        public async Task<ResponseManager> DeleteIssueAsync(IntIdViewModel model)
         {
             try
             {
@@ -205,27 +375,54 @@ namespace CampusCore.API.Services
                     };
                 }
 
-                _context.Issues.Remove(issue);
-                var result = await _context.SaveChangesAsync();
-
-                if (result > 0)
-                {
-                    return new ResponseManager
-                    {
-                        IsSuccess = true,
-                        Message = "Issue deleted successfully"
-                    };
-                }
-                else
+                //remove in junction table first
+                var submissionIssue = await _context.SubmissionIssues.FindAsync(model.Id);
+                if (submissionIssue == null)
                 {
                     return new ErrorResponseManager
                     {
                         IsSuccess = false,
-                        Message = "Issue deletion failed",
-                        Errors = new List<string> { "Error occurred while deleting the issue" }
+                        Message = "Issue not associated with any submission",
+                        Errors = new List<string> { "Issue with the specified ID is not in submission issue table" }
                     };
                 }
+                _context.SubmissionIssues.Remove(submissionIssue);
+                var res = await _context.SaveChangesAsync();
+
+                if (res > 0)
+                {
+                    //then remove issue in issues table
+                    _context.Issues.Remove(issue);
+                    var result = await _context.SaveChangesAsync();
+                    if (result > 0)
+                    {
+                        return new ResponseManager
+                        {
+                            Message = "Issue deleted successfully!",
+                            IsSuccess = true
+                        };
+                    }
+                    else
+                    {
+                        return new ErrorResponseManager
+                        {
+                            Message = "Issue is deleted but failed to update with submission issues table",
+                            IsSuccess = false,
+                            Errors = new List<string>() { "Error deleting issue in DB" }
+                        };
+                    }
+
+
+                }
+                return new ErrorResponseManager
+                {
+                    Message = "Issue cannot be deleted in submission issues",
+                    IsSuccess = false,
+                    Errors = new List<string>() { "Error deleting issue in DB" }
+                };
+
             }
+
             catch (Exception ex)
             {
                 return new ErrorResponseManager
